requirements.txt
biopython==1.77
cycler==0.10.0
future==0.18.2
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.48.0
utils==1.0.1
